better logging and coroutine handling
